Update Home_Alarm_Device_Type
Added "motion sensor" as a capability.  The alarm panel will now mimic a motion sensor.  The motion sensor will be active when the alarm panel goes to active.  The motion sensor will be inactive when the system is disarmed.  

Also added tiles for zones 7-10.  
Added additional clarifications on the section of code which can be customized.