Update and rename AD2SmartThings_v4_3_1.ino to AD2SmartThings_v4_4.ino
fixed one bug:
now reports if zone is active and system tries to arm

Addressed comments from mlritchie:
Line 11, You can remove mention of ArduinoJson as we no longer use it
done
Lines 60-63 - I see you removed the verbosity settings for debugging. I found that feature very useful when I was trying to find certain things but if you prefer to remove it, you should remove these lines too. You will also need to remove it from line 414 as well. The reason I put in the serialLog is that it reduced the number of lines of code as to debug you now have 3 lines as opposed to 1 line calling a function that evaluates isDebugEnabled in my version; something to consider.
my goal was to remove as much excess code as possible to minimize time out of loop. When out of loop(), the Arduino cannot respond to messages from alarm panel or from the device type.  Also, it just slows the response time before the Arduino gets around to the smartthing.send command.  Even when debug was turned off, it still had to evaluate multiple serialLog commands. The only time that the Arduino could evaluate a device handler command was the brief time between messages from the alarm panel.  The gap was very small.  There is now a big gap.
I actually liked your subroutine so I added it back in, just with fewer serialLog commands.  I think the average message only has to evaluate ~3-4 serialLog commands now.  Should be no big deal, even for the Arduino processor.  
I like the idea of checking to see if the string is the same at the beginning and just returning out of the loop.
With changes to Chime, I noticed some weird things happening on my panel where the message wasn't being passed which is why my version was forcing the keypad message. I see you removed that code so we will need to test this out.
It works beautifully!!
You really don't need the resetZoneStatusList function as the getActiveList function is doing the same thing except it returns a fault list. In the event you don't want the fault list, you just need to ignore the return value. Less code to maintain. You just need to call getActiveList like the following: getActiveList(1, numZones + 1);
great catch.  done.
Lines 318-321, what do you mean by this? You are clearing out zones that may still active. Sure once the panel posts the next message with the next higher zone, that zone will become active again, but this is making it active, inactive, and active again. If there are automations against these virtual devices, that could cause problems.
My bad on the annotation.  I updated annotation.  We can discuss on the phone
Around line 339, you may want to repeat the line you have in the device handler so people know there is a required order for the string:
0:powerStatus, 1:chimeStatus, 2:alarmStatus, 3:activeZone, 4:inactiveList, 5:keypadMsg
done