Update and rename Home_Alarm_Device_Type_v4_3 to Home_Alarm_Device_Type_v4_4
bug fixes:
fixed processing of !sending..done message
suppressed logging of keypad messages to the "Recently" log on the smart things app.

addressed comments from mlritchie:
You may want to consider moving the "none" from activeZone to say an "allclear" on inactiveList. I suggest that only because we could easily leverage the for loop in the inactiveList to set them all inactive. Plus we should put a check in there anyway to see if it is active before marking it inactive anyway. I am happy to help code this up if you like the idea.

	I like it —done