Update Home_Alarm_Device_Type_v4_4
1. Updated the instructions at the top for the zones.  Included a note that the zones can be reordered if desired as I did that by putting all my motions at the top and my smoke and glass breaks at the bottom given they never trip.
2. Reworked the allClear code.  Since we need to know the number of zones to clear and I don't want the user to have to enter it both in the DH and sketch, I now have the sketch sending the number of zones so we know the number to check for active and clear.
3. Reworked many of the function calls at the bottom as a result of reworking the messageCallout function in the Arduino sketch.  Separating values by a | and I noted the schema. 

Should we consider updating the tiles to be a scale of 2?  I ask because this would allow for smaller buttons if people want them to add additional buttons for say arm stay instant, etc.  I did this on my own DH.

I have zone9 by default and this sketch doesn't include it.  Should we add it?  Not sure how common having no zone 9 is because that is the first zone in my zone expander.