Update and rename Alarm_Handler_SmartApp to Alarm_Handler_SmartApp_v2
Added support for additional notifications on selected contact sensors and motions.  Use case is to be notified if an alarm device's state changes it will notify you.  My personal use case was me adding the new smoke and CO relays and I wanted to be notified when the state changes so I just added the functionality to this SmartApp.  Figured others could benefit as well.