Update Home_Alarm_Device_Type_v4_4_3
Found an extra debug message, removed it. Also update the processing of inactiveArray to NOT show keypad message since it is useful to know which zone is no longer active.  keypad message is great for active zones since it indicates the zone active but since inactive are never on the keypad message this is the only way to track specific inactive zone numbers.